Updated Importer for Updated CSV format. Added Recipe Level to Recipe. Added View Single Recipe Updated Recipe Search with Recipe Level.

+++ CraftingCalculator/DTOs/IngredientDTO.cs
@@ -3,7 +3,7 @@
 public record IngredientDTO
 {
     public uint Id { get; set; }
-    public uint Count { get; set; }
+    public byte Count { get; set; }
     public RecipeDTO RecipeDTO { get; set; }
     public ItemDTO ItemDTO { get; set; }
 }
