+++ CraftingCalculator/Models/Ingredient.cs
@@ -3,7 +3,7 @@
 public class Ingredient
 {
     public uint Id { get; set; }
-    public uint Count { get; set; }
+    public byte Count { get; set; }
     public uint RecipeId { get; set; }
     public Recipe Recipe { get; set; }
     public uint ItemId { get; set; }
